% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{anova.emaxlogistic}
\alias{anova.emaxlogistic}
\title{Likelihood ratio test for logistic Emax regression models}
\usage{
\method{anova}{emaxlogistic}(object, ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{...}{Additional fitted \code{emaxlogistic} model objects}
}
\value{
A data frame with columns \code{Df}, \code{Deviance}, \code{Df_diff}, \code{LRT}, and \verb{Pr(>Chi)}
}
\description{
Likelihood ratio test for logistic Emax regression models
}
\details{
Computes a likelihood ratio chi-squared test comparing a sequence of nested
logistic Emax models. The test statistic is the difference in deviances;
the reference distribution is chi-squared with degrees of freedom equal to
the difference in the number of parameters.
}
