% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{anova.emaxnls}
\alias{anova.emaxnls}
\title{Analysis of variance for Emax regression models}
\usage{
\method{anova}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Additional fitted model objects}
}
\value{
Analysis of variance tables for a sequence of \code{emaxnls} models
}
\description{
Analysis of variance for Emax regression models
}
\examples{

mod_0 <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ 1, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
mod_1 <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)

anova(mod_0, mod_1)

}
