% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{anova.emaxnls_fit}
\alias{anova.emaxnls_fit}
\title{Analysis of variance for Emax regression models}
\usage{
\method{anova}{emaxnls_fit}(object, ...)
}
\arguments{
\item{object}{An `emaxnls_fit`` object}

\item{...}{Additional fitted model objects}
}
\value{
Analysis of variance tables for a sequence of \code{emaxnls_fit} models
}
\description{
Analysis of variance for Emax regression models
}
