% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{BIC.emaxlogistic}
\alias{BIC.emaxlogistic}
\title{BIC for a logistic Emax regression}
\usage{
\method{BIC}{emaxlogistic}(object, ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{...}{Optionally more fitted model objects}
}
\value{
Numeric, or a data frame when multiple models are supplied
}
\description{
BIC for a logistic Emax regression
}
