% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.emax_fit}
\alias{coef.emax_fit}
\title{Coefficents for an Emax regression}
\usage{
\method{coef}{emax_fit}(object, level = 0.95, ...)
}
\arguments{
\item{object}{An emax_fit object}

\item{level}{Confidence level for interval estimate}

\item{...}{Ignored}
}
\value{
A tibble
}
\description{
Coefficents for an Emax regression
}
