% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{deviance.emaxlogistic}
\alias{deviance.emaxlogistic}
\title{Model deviance for a logistic Emax regression}
\usage{
\method{deviance}{emaxlogistic}(object, ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Model deviance for a logistic Emax regression
}
