% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{deviance.emaxnls}
\alias{deviance.emaxnls}
\title{Model deviance for an Emax regression}
\usage{
\method{deviance}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Model deviance for an Emax regression
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
deviance(mod)


}
