% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{df.residual.emaxlogistic}
\alias{df.residual.emaxlogistic}
\title{Residual degrees of freedom for a logistic Emax regression}
\usage{
\method{df.residual}{emaxlogistic}(object, ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Residual degrees of freedom for a logistic Emax regression
}
