% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{df.residual.emaxnls}
\alias{df.residual.emaxnls}
\title{Residual degrees of freedom for an Emax regression model}
\usage{
\method{df.residual}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Residual degrees of freedom for an Emax regression model
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
df.residual(mod)

}
