% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R, R/emaxnls-methods.R
\name{df.residual.emaxlogistic}
\alias{df.residual.emaxlogistic}
\alias{df.residual}
\alias{df.residual.emaxnls}
\title{Residual degrees of freedom for an Emax regression model}
\usage{
\method{df.residual}{emaxlogistic}(object, ...)

\method{df.residual}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} or \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
A numeric scalar
}
\description{
Returns the residual degrees of freedom, equal to the number of
observations minus the number of estimated parameters.
}
\details{
For \code{emaxnls} objects, the value is obtained directly from the underlying
\code{nls} fit. For \code{emaxlogistic} objects, it is computed as
\code{nobs(object) - length(coef(object))}.
}
\examples{
mod_c <- emax_nls(
  structural_model = rsp_1 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
df.residual(mod_c)

mod_b <- emax_logistic(
  structural_model = rsp_2 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
df.residual(mod_b)

}
