% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_auto_init}
\alias{emax_auto_init}
\title{Construct initial guess of a Emax model parameters}
\usage{
emax_auto_init(structural_model, covariate_model, data)
}
\arguments{
\item{structural_model}{A two-sided formula of the form response ~ exposure}

\item{covariate_model}{A list of two-sided formulas, each of specifying a covariate model for a structural parameter}

\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Construct initial guess of a Emax model parameters
}
\examples{
# use a heuristic to construct sensible start values, and plausible
# upper and lower bounds within which the estimate is expected to fall 
emax_auto_init(
  structural_model = response_1 ~ exposure_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)

# compare to the values estimated:
coef(emax_nls(
  structural_model = response_1 ~ exposure_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
))

}
