% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emax_df}
\alias{emax_df}
\title{Sample simulated data for Emax exposure-response models with covariates.}
\format{
A data frame with columns:
\describe{
\item{id}{Identifier column}
\item{dose}{Nominal dose, units not specified}
\item{exp_1}{Exposure value, units and metric not specified}
\item{exp_2}{Exposure value, units and metric not specified, but different from exp_1}
\item{rsp_1}{Continuous response value (units not specified)}
\item{rsp_2}{Binary response value (group labels not specified)}
\item{cnt_a}{Continuous valued covariate}
\item{cnt_b}{Continuous valued covariate}
\item{cnt_c}{Continuous valued covariate}
\item{bin_d}{Binary valued covariate}
\item{bin_e}{Binary valued covariate}
\item{cat_f}{Categorical covariate}
}
}
\usage{
emax_df
}
\description{
Sample simulated data for Emax exposure-response models with covariates.
}
\details{
This simulated dataset is entirely synthetic. It is a generic data set that can be used
to illustrate Emax modelling. It contains variables corresponding to dose and exposure,
and includes both a continuous response variable and a binary response variable. Three
continuous valued covariates are included, along with two binary covariates.

You can find the data generating code in the package source code,
under \code{R/data.R}
}
\examples{
emax_df
}
\keyword{datasets}
