% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_fn}
\alias{emax_fn}
\title{Construct Emax function from model object}
\usage{
emax_fn(mod)
}
\arguments{
\item{mod}{An \code{emaxnls} object}
}
\value{
A function with arguments \code{data} and \code{params}
}
\description{
Construct Emax function from model object
}
\examples{

mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
par <- coef(mod)

# customizable emax function with the same structural 
# model and same covariate model, defaulting to the 
# same data and parameters as the original model, but
# allowing user to pass their own data and parameters  
mod_fn <- emax_fn(mod)

# apply the function to a few rows of the original data
mod_fn(
  data = emax_df[120:125, ],
  param = par
)

# adjust the parameters
new_par <- par
new_par["E0_Intercept"] <- 0

# simulate the model with the adjusted parameters
mod_fn(
  data = emax_df[120:125, ],
  param = new_par
)

}
