% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_nls_settings}
\alias{emax_nls_settings}
\title{Settings used to estimate Emax model}
\usage{
emax_nls_settings(
  init = NULL,
  algorithm = "port",
  control = list(tol = 1e-08, minFactor = 1024^-4, maxiter = 2e+05, scaleOffset = 1,
    warnOnly = FALSE),
  ...
)
}
\arguments{
\item{init}{Data frame specifying initial parameters (start, upper, lower)}

\item{algorithm}{Has same meaning as in nls() (allowed: "default", "plinear", "port")}

\item{control}{Has same meaning as in nls()}

\item{...}{Other arguments passed to nls()}
}
\value{
List
}
\description{
Settings used to estimate Emax model
}
