% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{fitted.emaxlogistic}
\alias{fitted.emaxlogistic}
\title{Fitted values for a logistic Emax regression}
\usage{
\method{fitted}{emaxlogistic}(object, type = c("response", "link"), ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{type}{\code{"response"} (default) returns fitted probabilities;
\code{"link"} returns the linear predictor on the logit scale}

\item{...}{Ignored}
}
\value{
Numeric vector
}
\description{
Fitted values for a logistic Emax regression
}
