% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{fitted.emaxnls}
\alias{fitted.emaxnls}
\title{Fitted values for an Emax regression}
\usage{
\method{fitted}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric vector of fitted values
}
\description{
Fitted values for an Emax regression
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
fit <- fitted(mod)
fit[1:20]

}
