% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{logLik.emaxlogistic}
\alias{logLik.emaxlogistic}
\title{Log-likelihood for a logistic Emax regression}
\usage{
\method{logLik}{emaxlogistic}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{REML}{Ignored (included for S3 compatibility only)}

\item{...}{Ignored}
}
\value{
An object of class \code{logLik}
}
\description{
Log-likelihood for a logistic Emax regression
}
