% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{logLik.emaxnls}
\alias{logLik.emaxnls}
\title{Log-likelihood for an Emax regression model}
\usage{
\method{logLik}{emaxnls}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{REML}{For \code{nls} objects only \code{REML = FALSE} is supported}

\item{...}{Ignored}
}
\value{
Returns an object of class \code{logLik}. This is a number with
at least one attribute, "df" (degrees of freedom), giving the
number of (estimated) parameters in the model.
}
\description{
Log-likelihood for an Emax regression model
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
logLik(mod)

}
