% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R, R/emaxnls-methods.R
\name{logLik.emaxlogistic}
\alias{logLik.emaxlogistic}
\alias{logLik}
\alias{logLik.emaxnls}
\title{Log-likelihood for an Emax regression model}
\usage{
\method{logLik}{emaxlogistic}(object, REML = FALSE, ...)

\method{logLik}{emaxnls}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{An \code{emaxnls} or \code{emaxlogistic} object}

\item{REML}{For \code{emaxnls} objects only \code{REML = FALSE} is supported.
Ignored for \code{emaxlogistic} objects.}

\item{...}{Ignored}
}
\value{
An object of class \code{logLik} with at least one attribute, \code{"df"}
(degrees of freedom), giving the number of estimated parameters in the
model.
}
\description{
Log-likelihood for an Emax regression model
}
\examples{
mod_c <- emax_nls(
  structural_model = rsp_1 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
logLik(mod_c)

mod_b <- emax_logistic(
  structural_model = rsp_2 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
logLik(mod_b)

}
