% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{nobs.emaxnls}
\alias{nobs.emaxnls}
\title{Number of observations for an Emax regression model}
\usage{
\method{nobs}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Number of observations for an Emax regression model
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
nobs(mod)

}
