% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{predict.emaxnls}
\alias{predict.emaxnls}
\title{Predicting from Emax regression models}
\usage{
\method{predict}{emaxnls}(
  object,
  newdata = NULL,
  se.fit = FALSE,
  interval = "none",
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{newdata}{A named list or data frame in which to look for variables with which to predict.
If \code{newdata} is missing the fitted values at the original data points are returned.}

\item{se.fit}{A switch indicating if standard errors are required.}

\item{interval}{A character string indicating if prediction intervals or a confidence interval
on the mean responses are to be calculated. Can be "none", "confidence", or "prediction"}

\item{level}{A numeric scalar between 0 and 1 giving the confidence level for the intervals
(if any) to be calculated.}

\item{...}{Ignored}
}
\value{
The return value differs slightly depending on inputs. When \code{se.fit = FALSE}, it
produces a vector or matrix of predictions with column names \code{fit}, \code{lwr} and \code{upr} if
the \code{interval} argument is set. When \code{se.fit = TRUE}, it returns a list with the following
components:
\itemize{
\item \code{fit}: vector or matrix as above
\item \code{se.fit}: standard error of the predicted means
\item \code{residual.scale}: residual standard deviation
\item \code{df}: residual degrees of freedom
}
}
\description{
Predicting from Emax regression models
}
