% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.emax_fit}
\alias{print.emax_fit}
\title{Print an Emax regression model object}
\usage{
\method{print}{emax_fit}(x, ...)
}
\arguments{
\item{x}{An emax_fit object}

\item{...}{Ignored}
}
\value{
Invisibly returns the original object
}
\description{
Print an Emax regression model object
}
