% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-printing.R
\name{print.emaxlogistic}
\alias{print.emaxlogistic}
\title{Print a logistic Emax regression model object}
\usage{
\method{print}{emaxlogistic}(x, ...)
}
\arguments{
\item{x}{An \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
Invisibly returns the original object
}
\description{
Print a logistic Emax regression model object
}
