% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.emaxnls}
\alias{print.emaxnls}
\title{Print an Emax regression model object}
\usage{
\method{print}{emaxnls}(x, ...)
}
\arguments{
\item{x}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Invisibly returns the original object
}
\description{
Print an Emax regression model object
}
