% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-printing.R, R/emaxnls-printing.R
\name{print.emaxlogistic}
\alias{print.emaxlogistic}
\alias{print}
\alias{print.emaxnls}
\title{Print an Emax regression model object}
\usage{
\method{print}{emaxlogistic}(x, ...)

\method{print}{emaxnls}(x, ...)
}
\arguments{
\item{x}{An \code{emaxnls} or \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
Invisibly returns the original object
}
\description{
The \code{print()} method for \code{emaxnls} and \code{emaxlogistic} regression models
reports the structural and covariate model along with a table of
coefficients, test statistics, and confidence intervals
}
\examples{
mod_c <- emax_nls(
  structural_model = rsp_1 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
print(mod_c)

mod_b <- emax_logistic(
  structural_model = rsp_2 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
print(mod_b)

}
