% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.emax_fit}
\alias{residuals.emax_fit}
\title{Residuals for an Emax regression}
\usage{
\method{residuals}{emax_fit}(object, ...)
}
\arguments{
\item{object}{An emax_fit object}

\item{...}{Ignored}
}
\value{
Numeric vector of residuals
}
\description{
Residuals for an Emax regression
}
