% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{residuals.emaxlogistic}
\alias{residuals.emaxlogistic}
\title{Residuals for a logistic Emax regression}
\usage{
\method{residuals}{emaxlogistic}(object, type = c("pearson", "deviance"), ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{type}{\code{"pearson"} (default) or \code{"deviance"}}

\item{...}{Ignored}
}
\value{
Numeric vector
}
\description{
Residuals for a logistic Emax regression
}
