% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{residuals.emaxnls}
\alias{residuals.emaxnls}
\title{Residuals for an Emax regression}
\usage{
\method{residuals}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric vector of residuals
}
\description{
Residuals for an Emax regression
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
res <- residuals(mod)
res[1:20]
}
