% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R
\name{simulate.emaxlogistic}
\alias{simulate.emaxlogistic}
\title{Simulate responses from a logistic Emax regression model}
\usage{
\method{simulate}{emaxlogistic}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{nsim}{Number of replicates}

\item{seed}{Used to set RNG seed}

\item{...}{Ignored}
}
\value{
A data frame of simulated binary responses
}
\description{
Simulate responses from a logistic Emax regression model
}
\details{
Simulates new binary responses by (1) sampling parameter values from the
multivariate normal distribution implied by the estimated covariance matrix,
(2) computing predicted probabilities from each parameter draw, and (3)
drawing binary outcomes from \code{Bernoulli(p)} for each observation.
}
