% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{simulate.emaxnls}
\alias{simulate.emaxnls}
\title{Simulate responses from Emax regression model}
\usage{
\method{simulate}{emaxnls}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{nsim}{Number of replicates}

\item{seed}{Used to set RNG seed}

\item{...}{Ignored}
}
\value{
A data frame or tibble
}
\description{
Simulate responses from Emax regression model
}
\examples{
mod <- emax_nls(
  structural_model = response_1 ~ exposure_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
simulate(mod)

}
