% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-printing.R
\name{summary.emaxlogistic}
\alias{summary.emaxlogistic}
\title{Summary of a logistic Emax regression model}
\usage{
\method{summary}{emaxlogistic}(object, conf_level = 0.95, back_transform = FALSE, ...)
}
\arguments{
\item{object}{An \code{emaxlogistic} object}

\item{conf_level}{Confidence level for interval estimates}

\item{back_transform}{Should log-scaled parameters (logEC50, logHill) be
back-transformed to the original scale?}

\item{...}{Ignored}
}
\value{
A data frame containing a table of parameter estimates with
z-statistics, p-values, and confidence intervals
}
\description{
Summary of a logistic Emax regression model
}
