% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.emax_fit}
\alias{vcov.emax_fit}
\title{Variance-covariance matrix for an Emax regression}
\usage{
\method{vcov}{emax_fit}(object, ...)
}
\arguments{
\item{object}{An emax_fit object}

\item{...}{Ignored}
}
\value{
A matrix
}
\description{
Variance-covariance matrix for an Emax regression
}
