% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC}
\alias{AIC}
\alias{AIC.emaxnls_fit}
\alias{BIC.emaxnls_fit}
\title{Akaike information criterion / Bayesian information criterion}
\usage{
\method{AIC}{emaxnls_fit}(object, ..., k = 2)

\method{BIC}{emaxnls_fit}(object, ...)
}
\arguments{
\item{object}{An `emaxnls_fit`` object}

\item{...}{Optionally, more fitted model objects}

\item{k}{Penalty per parameter in the AIC}
}
\value{
If just one object is provided, a numeric value with the corresponding AIC (or BIC).
If multiple objects are provided, a data.frame with rows corresponding to the objects
and columns representing the number of parameters in the model (df) and the AIC or BIC.
}
\description{
Akaike information criterion / Bayesian information criterion
}
