% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{anova.emaxnls}
\alias{anova.emaxnls}
\title{Analysis of variance for Emax regression models}
\usage{
\method{anova}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Additional fitted model objects}
}
\value{
Analysis of variance tables for a sequence of \code{emaxnls} models
}
\description{
Analysis of variance for Emax regression models
}
