% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{coef.emaxnls}
\alias{coef.emaxnls}
\title{Coefficients for an Emax regression}
\usage{
\method{coef}{emaxnls}(object, back_transform = FALSE, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{back_transform}{Should log-scaled parameters (logEC50, logHill) be back-transformed to original scale?}

\item{...}{Ignored}
}
\value{
A vector of coefficients
}
\description{
Coefficients for an Emax regression
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)

# coefficients on the estimation scale
coef(mod)

# coefficients with log-scale parameters back-transformed
coef(mod, back_transform = TRUE)

}
