% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{coef.emaxnls}
\alias{coef.emaxnls}
\title{Coefficents for an Emax regression}
\usage{
\method{coef}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
A vector of coefficients
}
\description{
Coefficents for an Emax regression
}
