% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.emaxnls}
\alias{confint.emaxnls}
\title{Confidence intervals for Emax regression model parameters}
\usage{
\method{confint}{emaxnls}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{parm}{A specification of which parameters are to be given confidence intervals,
either a vector of numbers or a vector of names. If \code{parm = NULL}, all parameters are
considered.}

\item{level}{The confidence level required}

\item{...}{Ignored}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each
parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in \% (by default
2.5\% and 97.5\%).
}
\description{
Confidence intervals for Emax regression model parameters
}
