% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{deviance.emaxnls}
\alias{deviance.emaxnls}
\title{Model deviance for an Emax regression}
\usage{
\method{deviance}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Model deviance for an Emax regression
}
