% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxlogistic-methods.R, R/emaxnls-methods.R
\name{deviance.emaxlogistic}
\alias{deviance.emaxlogistic}
\alias{deviance}
\alias{deviance.emaxnls}
\title{Model deviance for an Emax regression model}
\usage{
\method{deviance}{emaxlogistic}(object, ...)

\method{deviance}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} or \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
Numeric. For \code{emaxnls} objects, returns the residual sum of
squares. For \code{emaxlogistic} objects, returns the binomial deviance
(\code{-2 * logLik}).
}
\description{
Model deviance for an Emax regression model
}
\examples{
# emaxnls deviance (residual sum of squares)
mod_c <- emax_nls(
  structural_model = rsp_1 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
deviance(mod_c)

# emaxlogistic deviance (binomial deviance)
mod_b <- emax_logistic(
  structural_model = rsp_2 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
deviance(mod_b)

}
