% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{df.residual.emaxnls}
\alias{df.residual.emaxnls}
\title{Residual degrees of freedom for an Emax regression model}
\usage{
\method{df.residual}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Residual degrees of freedom for an Emax regression model
}
