% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_converged}
\alias{emax_converged}
\title{Check Emax regression model for convergence status}
\usage{
emax_converged(mod)
}
\arguments{
\item{mod}{An \code{emaxnls} object}
}
\value{
Logical value
}
\description{
Check Emax regression model for convergence status
}
