% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_converged}
\alias{emax_converged}
\title{Check Emax regression model for convergence status}
\usage{
emax_converged(mod)
}
\arguments{
\item{mod}{An \code{emaxnls} object}
}
\value{
Logical value

This is a convenience function that takes an Emax regression
object as input. It returns \code{TRUE} if the optimization routine
converged during model fitting, and \code{FALSE} if it did not.
}
\description{
Check Emax regression model for convergence status
}
