% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_fun}
\alias{emax_fun}
\title{Construct Emax prediction function from model object}
\usage{
emax_fun(mod)
}
\arguments{
\item{mod}{An \code{emaxnls} object}
}
\value{
A function \code{f} with arguments \code{data} and \code{params}. The \code{data}
argument defaults to the data used to estimate the model, and the
\code{params} arugment defaults to the estimated parameter values. Both
can be customized, as long as \code{data} contains columns corresponding
to each of the variables used by the model, and \code{params} is a named
numeric vector of the appropriate length. The names for \code{params}
must exactly match the names of the vector returned by \code{coef(mod)}.

The return value for \code{f} is a numeric vector of model predictions for
each row in \code{data}, evaluated at parameters \code{params}.
}
\description{
Construct Emax prediction function from model object
}
\examples{

mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
par <- coef(mod)

# customizable emax function with the same structural 
# model and same covariate model, defaulting to the 
# same data and parameters as the original model, but
# allowing user to pass their own data and parameters  
mod_fn <- emax_fun(mod)

# apply the function to a few rows of the original data
mod_fn(
  data = emax_df[120:125, ],
  param = par
)

# adjust the parameters
new_par <- par
new_par["E0_Intercept"] <- 0

# simulate the model with the adjusted parameters
mod_fn(
  data = emax_df[120:125, ],
  param = new_par
)

}
