% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_nls}
\alias{emax_nls}
\title{Emax model with arbitrary covariates (does not support interactions)}
\usage{
emax_nls(structural_model, covariate_model, data, quiet = FALSE)
}
\arguments{
\item{structural_model}{A two-sided formula of the form response ~ exposure}

\item{covariate_model}{A list of two-sided formulas, each of specifying a covariate model for a structural parameter}

\item{data}{A data frame}

\item{quiet}{When quiet=TRUE messages and warnings are suppressed}
}
\value{
An object of class \code{emaxnls}
}
\description{
Emax model with arbitrary covariates (does not support interactions)
}
\examples{
emax_nls(
  structural_model = response_1 ~ exposure_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
 
}
