% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_nls}
\alias{emax_nls}
\title{Emax model with arbitrary covariates (does not support interactions)}
\usage{
emax_nls(structural_model, covariate_model, data, dosing = NULL, quiet = FALSE)
}
\arguments{
\item{structural_model}{A two-sided formula of the form response ~ exposure}

\item{covariate_model}{A list of two-sided formulas, each of specifying a}

\item{data}{A data frame}

\item{dosing}{Column specifying the dosing regimen}

\item{quiet}{When quiet=TRUE messages and warnings are suppressed}
}
\value{
An object of class \code{emax_fit}
}
\description{
Emax model with arbitrary covariates (does not support interactions)
}
