% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_nls_init}
\alias{emax_nls_init}
\title{Construct an initial guess for the Emax model parameters}
\usage{
emax_nls_init(structural_model, covariate_model, data)
}
\arguments{
\item{structural_model}{A two-sided formula of the form response ~ exposure}

\item{covariate_model}{A list of two-sided formulas, each of specifying a
covariate model for a structural parameter}

\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Construct an initial guess for the Emax model parameters
}
\details{
The \code{emax_nls()} function requires that the user specify the initial
values for the model parameters. Specifically, it expects to be supplied with
a data frame with columns named \code{parameter}, \code{covariate}, and \code{start}. If a
bounded optimization method is used (e.g. if the "port" method is used), the
data frame also needs to have columns named \code{lower} and \code{upper}. The data
frame should contain one row per parameter. In most cases the user does not
need to define this manually, because \code{emax_nls_init()} can use heuristics to
make a sensible guess about what to use as starting values. By default this
is what \code{emax_nls()} relies upon, automatically calling \code{emax_nls_init()}
using the appropriate values for the \code{structural_model}, the \code{covariate_model},
and the \code{data}.
}
\examples{
# use a heuristic to construct sensible start values, and plausible
# upper and lower bounds within which the estimate is expected to fall 
emax_nls_init(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)

# compare to the values estimated:
coef(emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
))

}
\seealso{
\code{emax_nls()}, \code{emax_nls_options()}
}
