% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_nls_options}
\alias{emax_nls_options}
\title{Settings used to estimate Emax model}
\usage{
emax_nls_options(
  optim_method = "port",
  optim_control = NULL,
  quiet = FALSE,
  weights = NULL,
  na.action = options("na.action")
)
}
\arguments{
\item{optim_method}{Character string specifying the algorithm used to solve
the nonlinear least squares optimization problem. Supported pptions are
"gauss" (the default), "golub", "port", and "levenberg". See details.}

\item{optim_control}{A list of arguments used to control the behavior of
the optimization algorithm. Allowed values differ depending on which
algorithm is used}

\item{quiet}{When \code{quiet=TRUE}, messages are suppressed}

\item{weights}{Numeric vector providing the weights for observations. When
specified, weighted least squares is used}

\item{na.action}{How should missing values in the data be handled

There are four supported values for \code{optim_method}:
\itemize{
\item "gauss": The Gauss-Newton algorithm, and equivalent to the "default" option
in \code{nls()}
\item "golub": The Golub-Pereyra algorithm for partially linear least-squares,
equivalent to using "plinear" in \code{nls()}
\item "port": The "nl2sol" algorithm from from the the Port library, equivalent
to "port" in \code{nls()}
\item "levenberg": The Levenberg-Marquardt algorithm. Equivalent to \code{nlsLM()}
from the "minpack.lm" package. To use this method, the "minpack.lm"
package must be installed
}}
}
\value{
List
}
\description{
Settings used to estimate Emax model
}
