% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_nls_options}
\alias{emax_nls_options}
\title{Settings used to estimate Emax model}
\usage{
emax_nls_options(
  optim_method = "port",
  optim_control = NULL,
  quiet = FALSE,
  weights = NULL,
  na.action = options("na.action")
)
}
\arguments{
\item{optim_method}{Character string specifying the algorithm used to solve
the nonlinear least squares optimization problem. Supported pptions are
"gauss", "port", and "levenberg". See details.}

\item{optim_control}{A list of arguments used to control the behavior of
the optimization algorithm. Allowed values differ depending on which
algorithm is used}

\item{quiet}{When \code{quiet=TRUE}, messages are suppressed}

\item{weights}{Numeric vector providing the weights for observations. When
specified, weighted least squares is used}

\item{na.action}{How should missing values in the data be handled

At present there are three supported values for \code{optim_method}:
\itemize{
\item "gauss": Estimate parameters using the Gauss-Newton algorithm. This is
equivalent to the using "default" option in \code{nls()}
\item "port": Estimate parameters using bounded optimization with the "nl2sol"
algorithm from from the the Port library. Equivalent to "port" in \code{nls()}
\item "levenberg": Estimate parameters using the Levenberg-Marquardt algorithm.
This is equivalent to using \code{nlsLM()} from the "minpack.lm" package.
}

Note that the Golub-Pereyra algorithm for partially linear least-squares (i.e. the
"plinear" option in \code{nls()}) is not currently supported for Emax regression. Informal
testing suggests it does not perform well for these models, and rarely converges.}
}
\value{
List
}
\description{
Settings used to estimate Emax model
}
