% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_scm}
\alias{emax_scm}
\alias{emax_scm_forward}
\alias{emax_scm_backward}
\alias{emax_scm_history}
\title{Stepwise covariate modeling for Emax regression}
\usage{
emax_scm_forward(mod, candidates, threshold = 0.01, seed = NULL)

emax_scm_backward(mod, candidates, threshold = 0.001, seed = NULL)

emax_scm_history(mod)
}
\arguments{
\item{mod}{An \code{emaxnls} object}

\item{candidates}{A list of candidate covariates}

\item{threshold}{Threshold for addition or removal}

\item{seed}{Seed for the RNG state}
}
\value{
An object of class \code{emaxnls}
}
\description{
Stepwise covariate modeling for Emax regression
}
\details{
The emaxnls package supports stepwise covariate modelling via forward addition and
backward elimination. The \code{emax_scm_forward()} function supports forward addition,
the \code{emax_scm_backward()} function supports backward elimination, and the syntax
is designed to allow forward-backward procedures by piping a base model to
\code{emax_scm_forward()} and then to \code{emax_scm_backward()}. In both cases, the
function takes an \code{emaxnls} regression object as the first argument, as well as
a list of candidate \code{covariates} to be considered for addition (in the forward
addition case) or deletion (backward elimination). The input must be a named list,
with the names corresponding to the relevant structural parameter, and the values
should be character vector specifying covariates of interest. See the examples for
an illustration of how this argument should be specified.

As present, these functions only support stepwise regression using p-values as the
criterion for addition or deletion. The \code{threshold} argument corresponds to the
threshold p-value to be used. In future, other methods (e.g., selection on the
basis of AIC values) may be supported.

The \code{seed} argument is used to control the RNG state for stochastic components of
the stepwise procedure. However, please note that the \code{seed} argument is currently
experimental, and may be removed in future releases.

A key feature of the stepwise covariate modelling functions is that they keep track
of every tested model, and store information about this history internally within the
\code{emaxnls} object that gets returned. Use the \code{emax_scm_history()} function to extract
this history.
}
\examples{
base_model <- emax_nls(rsp_1 ~ exp_1, list(E0 ~ 1, Emax ~ 1, logEC50 ~ 1), emax_df)

covariate_list <- list(
  E0 = c("cnt_a", "cnt_b", "cnt_c", "bin_d", "bin_e"),
  Emax = c("cnt_a", "cnt_b", "cnt_c", "bin_d", "bin_e")
)

# add covariates to the base model using forward addition
forward_model <- emax_scm_forward(
  mod = base_model,
  candidates = covariate_list, 
  threshold = .01
)
forward_model

# remove covariates from the forward model using backward deletion
final_model <- emax_scm_backward(
  mod = forward_model,
  candidates = covariate_list, 
  threshold = .001
) 
final_model

# show the history of all models tested during the forward addition
# step and the backward deletion step
emax_scm_history(final_model)

}
