% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_update}
\alias{emax_update}
\alias{emax_add_term}
\alias{emax_remove_term}
\title{Add or remove a covariate term from an Emax regression}
\usage{
emax_add_term(mod, formula)

emax_remove_term(mod, formula)
}
\arguments{
\item{mod}{An \code{emaxnls} object}

\item{formula}{A formula such as E0 ~ AGE}
}
\value{
An object of class \code{emaxnls}
}
\description{
Add or remove a covariate term from an Emax regression
}
\details{
The \code{emax_add_term()} and \code{emax_remove_term()} functions take an existing Emax regression
object, and allow the user to add or remove a specific term to the model. It is not expected
that users will need these functions very often, but they provide the basis for the stepwise
covariate modeling procedures that are very commonly used when building Emax regressions.
}
\examples{
mod_0 <- emax_nls(rsp_1 ~ exp_1, list(E0 ~ 1, Emax ~ 1, logEC50 ~ 1), emax_df)
mod_1 <- emax_nls(rsp_1 ~ exp_1, list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), emax_df)

if (emax_converged(mod_0)) emax_add_term(mod_0, E0 ~ cnt_a)

if (emax_converged(mod_1)) emax_remove_term(mod_1, E0 ~ cnt_a)

}
\seealso{
\code{emax_nls()}, \link{emax_scm}
}
