% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{emax_update}
\alias{emax_update}
\alias{emax_add_term}
\alias{emax_remove_term}
\title{Add or remove a covariate term from an Emax regression}
\usage{
emax_add_term(object, formula, quiet = TRUE)

emax_remove_term(object, formula, quiet = TRUE)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{formula}{A formula such as E0 ~ AGE}

\item{quiet}{When quiet=TRUE messages and warnings are suppressed}
}
\value{
An object of class \code{emaxnls}
}
\description{
Add or remove a covariate term from an Emax regression
}
\examples{
mod_0 <- emax_nls(response_1 ~ exposure_1, list(E0 ~ 1, Emax ~ 1, logEC50 ~ 1), emax_df)
mod_1 <- emax_nls(response_1 ~ exposure_1, list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), emax_df)

emax_add_term(mod_0, E0 ~ cnt_a)
emax_remove_term(mod_1, E0 ~ cnt_a)

}
