% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{fitted.emaxnls}
\alias{fitted.emaxnls}
\title{Fitted values for an Emax regression}
\usage{
\method{fitted}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric vector of fitted values
}
\description{
Fitted values for an Emax regression
}
