% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.emaxnls}
\alias{logLik.emaxnls}
\title{Log-likelihood for an Emax regression model}
\usage{
\method{logLik}{emaxnls}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{REML}{For \code{nls} objects only \code{REML = FALSE} is supported}

\item{...}{Ignored}
}
\value{
Returns an object of class \code{logLik}. This is a number with
at least one attribute, "df" (degrees of freedom), giving the
number of (estimated) parameters in the model.
}
\description{
Log-likelihood for an Emax regression model
}
