% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{nobs.emaxnls}
\alias{nobs.emaxnls}
\title{Number of observations for an Emax regression model}
\usage{
\method{nobs}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} or \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
A numeric scalar
}
\description{
Returns the number of observations used when fitting the model.
}
\details{
This reflects the actual number of rows passed to the fitting algorithm
after any missing-value handling specified via the \code{na.action} option in
\code{emax_nls_options()} or \code{emax_logistic_options()}. The value is used
internally by \code{BIC()} and \code{df.residual()}.
}
\examples{
mod_c <- emax_nls(
  structural_model = rsp_1 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df,
  opts = emax_nls_options(max_time = 10)
)
nobs(mod_c)

mod_b <- emax_logistic(
  structural_model = rsp_2 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df,
  opts = emax_logistic_options(max_time = 10)
)
nobs(mod_b)

}
