% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{nobs.emaxnls}
\alias{nobs.emaxnls}
\title{Number of observations for an Emax regression model}
\usage{
\method{nobs}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Number of observations for an Emax regression model
}
