% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{predict.emaxnls}
\alias{predict.emaxnls}
\title{Predicting from Emax regression models}
\usage{
\method{predict}{emaxnls}(
  object,
  newdata = NULL,
  se.fit = FALSE,
  interval = "none",
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{newdata}{A named list or data frame in which to look for variables with which to predict.
If \code{newdata} is missing the fitted values at the original data points are returned.}

\item{se.fit}{A switch indicating if standard errors are required.}

\item{interval}{A character string indicating if prediction intervals or a confidence interval
on the mean responses are to be calculated. Can be "none", "confidence", or "prediction"}

\item{level}{A numeric scalar between 0 and 1 giving the confidence level for the intervals
(if any) to be calculated.}

\item{...}{Ignored}
}
\value{
As \code{xgxr::predict.nls()}.
}
\description{
Predicting from Emax regression models
}
\details{
The \code{predict()} method for for Emax regression is a thin wrapper around \code{xgxr::predict.nls()}.
Please see the documentation for that function.
}
