% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.emaxnls}
\alias{predict.emaxnls}
\title{Predicting from Emax regression models}
\usage{
\method{predict}{emaxnls}(object, newdata = NULL, interval = "none", level = 0.95, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{newdata}{A named list or data frame in which to look for variables with which to predict.
If \code{newdata} is missing the fitted values at the original data points are returned.}

\item{interval}{A character string indicating if prediction intervals or a confidence interval
on the mean responses are to be calculated. Can be "none", "confidence", or "prediction"}

\item{level}{A numeric scalar between 0 and 1 giving the confidence level for the intervals
(if any) to be calculated.}

\item{...}{Ignored}
}
\value{
The \code{predict()} method for Emax regression objects always returns a data frame or tibble
}
\description{
Predicting from Emax regression models
}
\details{
Note that \code{predict()} for Emax regression calls \code{investr::predFit()} to obtain confidence and
prediction intervals. Please see the documentation for that function.
}
