% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.emaxnls}
\alias{residuals.emaxnls}
\title{Residuals for an Emax regression}
\usage{
\method{residuals}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric vector of residuals
}
\description{
Residuals for an Emax regression
}
