% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sigma.emaxnls}
\alias{sigma.emaxnls}
\title{Residual standard deviation for Emax regression models}
\usage{
\method{sigma}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
Numeric
}
\description{
Residual standard deviation for Emax regression models
}
