% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{summary.emaxnls}
\alias{summary.emaxnls}
\title{Summary of an Emax regression model}
\usage{
\method{summary}{emaxnls}(object, conf_level = 0.95, back_transform = FALSE, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{conf_level}{Confidence level for interval estimates}

\item{back_transform}{Should log-scaled parameters (logEC50, logHill) be back-transformed to original scale?}

\item{...}{Ignored}
}
\value{
A data frame or tibble containing a table of parameter estimates and other statistical summaries.
Please note that the \code{summary()} method is experimental (moreso than other methods), and the return value
may be modified in future releases as the package matures.
}
\description{
Summary of an Emax regression model
}
\examples{
mod <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)

# standard summary
summary(mod)

# summary with adjusted confidence level
summary(mod, conf_level = 0.99)

# summary with log-scale parameters transformed to original scale
summary(mod, back_transform = TRUE)

}
