% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxnls-methods.R
\name{vcov.emaxnls}
\alias{vcov.emaxnls}
\title{Variance-covariance matrix for an Emax regression}
\usage{
\method{vcov}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} or \code{emaxlogistic} object}

\item{...}{Ignored}
}
\value{
A matrix
}
\description{
Variance-covariance matrix for an Emax regression
}
\examples{
mod_c <- emax_nls(
  structural_model = rsp_1 ~ exp_1, 
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1), 
  data = emax_df
)
vcov(mod_c)

mod_b <- emax_logistic(
  structural_model = rsp_2 ~ exp_1,
  covariate_model = list(E0 ~ cnt_a, Emax ~ 1, logEC50 ~ 1),
  data = emax_df
)
vcov(mod_b)

}
