% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.emaxnls}
\alias{vcov.emaxnls}
\title{Variance-covariance matrix for an Emax regression}
\usage{
\method{vcov}{emaxnls}(object, ...)
}
\arguments{
\item{object}{An \code{emaxnls} object}

\item{...}{Ignored}
}
\value{
A matrix
}
\description{
Variance-covariance matrix for an Emax regression
}
